+++ pydoclint/utils/return_yield_raise.py
@@ -1,5 +1,5 @@
 import ast
-from typing import Tuple, Type, Union
+from typing import Dict, Tuple, Type, Union
 
 from pydoclint.utils import walk
 from pydoclint.utils.annotation import unparseAnnotation
@@ -10,6 +10,7 @@ ReturnType = Type[ast.Return]
 ExprType = Type[ast.Expr]
 YieldAndYieldFromTypes = Tuple[Type[ast.Yield], Type[ast.YieldFrom]]
 FuncOrAsyncFuncTypes = Tuple[Type[ast.FunctionDef], Type[ast.AsyncFunctionDef]]
+FuncOrAsyncFunc = (ast.FunctionDef, ast.AsyncFunctionDef)
 
 
 def hasReturnAnnotation(node: FuncOrAsyncFuncDef) -> bool:
@@ -45,20 +46,37 @@ def hasIteratorOrIterableAsReturnAnnotation(node: FuncOrAsyncFuncDef) -> bool:
 
 def hasYieldStatements(node: FuncOrAsyncFuncDef) -> bool:
     """Check whether the function node has any yield statements"""
-    thisId = getFunctionId(node)
+    childLine: int = -999
+    foundYieldStmt: bool = False
+
+    # key: child lineno, value: (parent lineno, is parent a function?)
+    familyTree: Dict[int, Tuple[int, bool]] = {}
+
     for child, parent in walk.walk(node):
+        childLine = _getLineNum(child)
+        parentLine = _getLineNum(parent)
+        if childLine != -1 and parentLine != -1 and childLine != parentLine:
+            isFunction = isinstance(parent, FuncOrAsyncFunc)
+            familyTree[childLine] = (parentLine, isFunction)
+
         if isinstance(child, ast.Expr) and isinstance(
             child.value, (ast.Yield, ast.YieldFrom)
         ):
             if isinstance(parent, (ast.AsyncFunctionDef, ast.FunctionDef)):
-                # Only return True if the parent is `node` (in other words,
-                # this yield statement doesn't come from a nested function)
-                parentId = getFunctionId(parent)
-                if thisId == parentId:
-                    return True
+                foundYieldStmt = True
+                break
 
             if isinstance(parent, BlockType):
-                return True
+                foundYieldStmt = True
+                break
+
+    if foundYieldStmt:
+        parentFuncLineNum = _lookupParentFunc(familyTree, childLine)
+
+        # We consider this `yield` a valid one only when its parent function
+        # is indeed `node`
+        if parentFuncLineNum == node.lineno:
+            return True
 
     return False
 
@@ -95,3 +113,27 @@ def _hasReturnOrRaiseStatements(
 
 def _isNone(node: ast.AST) -> bool:
     return isinstance(node, ast.Constant) and node.value is None
+
+
+def _getLineNum(node: ast.AST) -> int:
+    try:
+        lineNum = node.lineno
+    except Exception:
+        lineNum = -1
+
+    return lineNum
+
+
+def _lookupParentFunc(
+        familyLine: Dict[int, Tuple[int, bool]],
+        lineNum: int,
+) -> int:
+    if lineNum not in familyLine:
+        return -999
+
+    parentLineNum, isParentAFunction = familyLine[lineNum]
+
+    if isParentAFunction:
+        return parentLineNum
+
+    return _lookupParentFunc(familyLine, parentLineNum)